+++ src/components/Loans/OpenLoans/components/Modals/Modals.js
@@ -6,6 +6,7 @@ import { stripesShape } from '@folio/stripes/core';
 
 // eslint-disable-next-line
 import PatronBlockModal from '@folio/users/src/components/PatronBlock/PatronBlockModal';
+import BulkClaimedReturnedModal from '../BulkClaimReturnedModal';
 
 class Modals extends React.Component {
   static propTypes = {
@@ -20,6 +21,9 @@ class Modals extends React.Component {
     hideChangeDueDateDialog: PropTypes.func.isRequired,
     changeDueDateDialogOpen: PropTypes.bool.isRequired,
     onClosePatronBlockedModal: PropTypes.func.isRequired,
+    requestCounts: PropTypes.object.isRequired,
+    onBulkClaimReturnedCancel: PropTypes.func.isRequired,
+    showBulkClaimReturnedModal: PropTypes.bool.isRequired,
   };
 
   static defaultProps = {
@@ -31,6 +35,7 @@ class Modals extends React.Component {
 
     const { stripes } = props;
     this.connectedChangeDueDateDialog = stripes.connect(ChangeDueDateDialog);
+    this.connectedBulkClaimReturnedDialog = stripes.connect(BulkClaimedReturnedModal);
   }
 
   render() {
@@ -46,12 +51,14 @@ class Modals extends React.Component {
       patronBlockedModal,
       onClosePatronBlockedModal,
       patronGroup,
+      requestCounts,
+      showBulkClaimReturnedModal,
+      onBulkClaimReturnedCancel,
     } = this.props;
 
     const loanIds = activeLoan
       ? loans.filter(loan => activeLoan === loan.id) // Only changing one due date.
       : loans.filter(loan => checkedLoans[loan.id]); // Bulk-changing due dates.
-
 
     return (
       <>
@@ -69,6 +76,12 @@ class Modals extends React.Component {
           patronBlocks={patronBlocks}
           onClose={onClosePatronBlockedModal}
           viewUserPath={`/users/view/${(user || {}).id}?filters=pg.${patronGroup.group}&sort=name`}
+        />
+        <this.connectedBulkClaimReturnedDialog
+          checkedLoansIndex={checkedLoans}
+          requestCounts={requestCounts}
+          open={showBulkClaimReturnedModal}
+          onCancel={onBulkClaimReturnedCancel}
         />
       </>
     );