+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * Prevent the renewal of claimed returned items. Refs UIU-1261.
 * Adds `checkedInFoundByLibrary` and `checkedInReturnedByPatron` to loan action map. Fixes UIU-1643.
 * Display the correct link for create/reset password. Refs UIU-1608.
+* Add a bulk claim returned function. Refs UIU-1627.
 * Format action message for claim returned checkin. Refs UIU-1218.
 * Capitalize user status when displayed. Fixes UIU-1523.